requirements.txt
admin==0.0.1
asgiref==3.2.10
bcrypt==3.2.0
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Django==3.0.8
django-rest-framework==0.1.0
django-storages==1.10
djangorestframework==3.11.0
docutils==0.15.2
firebase==3.0.1
gcloud==0.17.0
google-api-core==1.21.0
google-auth==1.19.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
jws==0.1.3
oauth2client==3.0.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
